feat(repo, internal/api, service): Implement client DeleteById. Refactor GetAll (added offset & limit params)

+++ src/internal/api/client/request/get_all.go
@@ -1,6 +1,6 @@
 package request
 
 type GetAllRequest struct {
-	Limit  uint64 `json:"limit,omitempty"`
-	Offset uint64 `json:"offset,omitempty"`
+	Limit  int `json:"limit,omitempty"`
+	Offset int `json:"offset,omitempty"`
 }
